Conductor-directed Option 1: bring QA1 current to the DR2-foundation base, then
merge AT-217 clean — preserving every lane. Union result carries:
  • QA1's 46 QA-only commits (AT-220 session armour, AT-211/213/214/221, AT-204/203,
    AT-200/201/199/192/197/194/164, …) — untouched, still walkable on qa1.
  • main's Staging lineage the foundation was cut from (~106 commits).
  • the DR2 shell (AT-215) + DR2 capture screen (AT-217, aa3649ef).

Conflicts resolved (base-divergence only, no logic lost):
  • routes/web.php — kept BOTH QA1's AT-220 /csrf-token and main's AT-178/AT-212
    reminders routes (additive); DR2 group auto-merged in.
  • PropertyWizardController.php + PropertyWizardDraftResumeTest.php — kept QA1's
    AT-211 resolveDraftAttribution() (newer than main's pre-AT-211 agent stamping)
    + its owner-attribution tests.
  • deals-v2/create-form.blade.php (sunset module) — took main's current lineage
    (AT-192d branch fix present on both lanes; nothing lost).
  • .ai/CHAT_STARTER.md + AT-192 audit — took main's newer docs.

Both lanes verified present: DR2 controller/view/routes/nav + AT-218 perm, and
AT-220 corex-session-guard.js + /csrf-token route. php -l clean on all resolved
files. Staging is the ceiling.

+++ app/Http/Resources/WebsiteApi/ListingResource.php
@@ -69,7 +69,10 @@ class ListingResource extends JsonResource
 
             // Dimensions / rooms.
             'beds'    => $this->beds !== null ? (int) $this->beds : null,
-            'baths'   => $this->baths !== null ? (float) $this->baths : null,
+            // Whole full-bath count. Kept an int (not float): the half surfaces
+            // separately in `half_baths`, and JSON can't carry a trailing `.0`
+            // anyway (json_encode collapses 1.0 -> 1 without PRESERVE_ZERO_FRACTION).
+            'baths'   => $this->baths !== null ? (int) $this->baths : null,
             // Half bathroom (guest toilet / cloakroom) count — a first-class
             // scalar alongside `baths`, NOT rolled into it. The website renders
             // total baths exactly as CoreX does everywhere: `baths + ½` when
